Create "game" module and migrate the domain part in.
