+++ game/src/main/java/tw/waterball/survive/io/game/sprite/bullet/Bullet.java
@@ -1,0 +1,17 @@
+package tw.waterball.survive.io.game.sprite.bullet;
+
+import tw.waterball.survive.io.game.sprite.Sprite;
+
+public class Bullet extends Sprite {
+    private int damage;
+    private int shootingRange;
+
+    public Bullet(int bulletWidth, int bulletHeight) {
+        super(bulletWidth, bulletHeight);
+    }
+
+    @Override
+    public void onUpdate() {
+
+    }
+}
