+++ game/src/main/java/tw/waterball/survive/io/game/sprite/role/Monster.java
@@ -1,0 +1,13 @@
+package tw.waterball.survive.io.game.sprite.role;
+
+public class Monster extends Role {
+
+    public Monster(int weaponSlotsSize, int monsterWidth, int monsterHeight) {
+        super(weaponSlotsSize, monsterWidth, monsterHeight);
+    }
+
+    @Override
+    public void onUpdate() {
+
+    }
+}
