+++ game/src/main/java/tw/waterball/survive/io/game/sprite/weapon/DroppedWeapon.java
@@ -1,0 +1,16 @@
+package tw.waterball.survive.io.game.sprite.weapon;
+
+/**
+ * @author - a89010531111@gmail.com
+ */
+public class DroppedWeapon {
+    private final Weapon weapon;
+
+    public DroppedWeapon(Weapon weapon) {
+        this.weapon = weapon;
+    }
+
+    public Weapon getWeapon() {
+        return weapon;
+    }
+}
