+++ game/src/main/java/tw/waterball/survive/io/game/sprite/weapon/Magazine.java
@@ -1,0 +1,11 @@
+package tw.waterball.survive.io.game.sprite.weapon;
+
+public class Magazine {
+    private int remainingBullets;
+    private int capacity;
+
+    public Magazine(int remainingBullets, int capacity) {
+        this.remainingBullets = remainingBullets;
+        this.capacity = capacity;
+    }
+}
