+++ game/src/main/java/tw/waterball/survive/io/game/sprite/weapon/WeaponName.java
@@ -1,0 +1,5 @@
+package tw.waterball.survive.io.game.sprite.weapon;
+
+public enum WeaponName {
+
+}
