+++ game/src/main/java/tw/waterball/survive/io/game/sprite/weapon/WeaponSlot.java
@@ -1,0 +1,14 @@
+package tw.waterball.survive.io.game.sprite.weapon;
+
+
+public class WeaponSlot {
+    private Weapon weapon;
+
+    public Weapon getWeapon() {
+        return weapon;
+    }
+
+    public void setWeapon(Weapon weapon) {
+        this.weapon = weapon;
+    }
+}
