+++ game/src/main/java/tw/waterball/survive/io/game/world/Event.java
@@ -1,0 +1,5 @@
+package tw.waterball.survive.io.game.world;
+
+public enum Event {
+
+}
