+++ game/src/main/java/tw/waterball/survive/io/game/world/GameLoop.java
@@ -1,0 +1,8 @@
+package tw.waterball.survive.io.game.world;
+
+/**
+ * @author - a89010531111@gmail.com
+ */
+public class GameLoop {
+    private SpriteContainer spriteContainer;
+}
