+++ pom.xml
@@ -26,6 +26,7 @@
         <module>server</module>
         <module>api</module>
         <module>ioc</module>
+        <module>game</module>
     </modules>
     <dependencies>
         <dependency>
