--- server/src/main/java/tw/waterball/survive/io/server/prototypefactory/WorldObjectName.java
@@ -1,5 +1,0 @@
-package tw.waterball.survive.io.server.prototypefactory;
-
-public enum WorldObjectName {
-
-}
