--- server/src/main/java/tw/waterball/survive/io/server/prototypefactory/WorldObjectPrototypeFactory.java
@@ -1,5 +1,0 @@
-package tw.waterball.survive.io.server.prototypefactory;
-
-public class WorldObjectPrototypeFactory {
-
-}
