--- server/src/main/java/tw/waterball/survive/io/server/world/Event.java
@@ -1,5 +1,0 @@
-package tw.waterball.survive.io.server.world;
-
-public enum Event {
-
-}
