--- server/src/main/java/tw/waterball/survive/io/server/worldobject/role/Monster.java
@@ -1,8 +1,0 @@
-package tw.waterball.survive.io.server.worldobject.role;
-
-public class Monster extends Role {
-
-    public Monster(int weaponSlotsSize, int monsterWidth, int monsterHeight) {
-        super(weaponSlotsSize, monsterWidth, monsterHeight);
-    }
-}
