--- server/src/main/java/tw/waterball/survive/io/server/worldobject/weapon/DroppedWeapon.java
@@ -1,16 +1,0 @@
-package tw.waterball.survive.io.server.worldobject.weapon;
-
-/**
- * @author - a89010531111@gmail.com
- */
-public class DroppedWeapon {
-    private final Weapon weapon;
-
-    public DroppedWeapon(Weapon weapon) {
-        this.weapon = weapon;
-    }
-
-    public Weapon getWeapon() {
-        return weapon;
-    }
-}
