--- server/src/main/java/tw/waterball/survive/io/server/worldobject/weapon/Magazine.java
@@ -1,11 +1,0 @@
-package tw.waterball.survive.io.server.worldobject.weapon;
-
-public class Magazine {
-    private int remainingBullets;
-    private int capacity;
-
-    public Magazine(int remainingBullets, int capacity) {
-        this.remainingBullets = remainingBullets;
-        this.capacity = capacity;
-    }
-}
