--- server/src/main/java/tw/waterball/survive/io/server/worldobject/weapon/Weapon.java
@@ -1,16 +1,0 @@
-package tw.waterball.survive.io.server.worldobject.weapon;
-
-import tw.waterball.survive.io.server.worldobject.WorldObject;
-import tw.waterball.survive.io.server.worldobject.role.Role;
-
-public abstract class Weapon extends WorldObject {
-    private WeaponName weaponName;
-    private Role weaponHolder;
-    private int damage;
-
-    public Weapon(int weaponWidth, int weaponHeight) {
-        super(weaponWidth, weaponHeight);
-    }
-
-    public abstract void useWeapon();
-}
