--- server/src/main/java/tw/waterball/survive/io/server/worldobject/weapon/WeaponName.java
@@ -1,5 +1,0 @@
-package tw.waterball.survive.io.server.worldobject.weapon;
-
-public enum WeaponName {
-
-}
