--- server/src/main/java/tw/waterball/survive/io/server/worldobject/weapon/WeaponSlot.java
@@ -1,14 +1,0 @@
-package tw.waterball.survive.io.server.worldobject.weapon;
-
-
-public class WeaponSlot {
-    private Weapon weapon;
-
-    public Weapon getWeapon() {
-        return weapon;
-    }
-
-    public void setWeapon(Weapon weapon) {
-        this.weapon = weapon;
-    }
-}
